[UDP] userIdx Int->Long

+++ app/src/main/java/com/example/swith/data/LoginData.kt
@@ -1,6 +1,6 @@
 package com.example.swith.data
 
 data class LoginData(
-    var userIdx:Int,
+    var userIdx:Long,
     var jwt: String,
 )
